Add utility getDocClientNewExpressionCount
